requirements.txt
discord.py
python-dotenv
pydub
pyaudio
